Implement Fiber Async, Guzzle Async, and Synchronous API Fetching with Performance Metrics
- Added index-php.php for handling API requests using Fiber Async, Guzzle, and synchronous methods.
- Implemented performance benchmarking and result display for each method.
- Created index.css for styling the results and loading indicators.
- Developed index.js for JavaScript-based async and sync API fetching with performance comparison.
- Included error handling and memory usage tracking for all methods.
- Enhanced user interface with loading animations and performance metrics display.
