+++ index.css
@@ -1,0 +1,194 @@
+body {
+    font-family: Arial, sans-serif;
+    max-width: 1400px;
+    margin: 0 auto;
+    padding: 20px;
+    background-color: #f5f5f5;
+}
+
+.guzzle-button {
+    background: #6f42c1;
+}
+
+.guzzle-button:hover {
+    background: #5a2d91;
+}
+
+.js-button {
+    background: #f7df1e;
+    color: #000;
+}
+
+.js-button:hover {
+    background: #e6c914;
+}
+
+.comparison-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.container {
+    background: white;
+    padding: 30px;
+    border-radius: 10px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.benchmark {
+    background: #e8f5e8;
+    padding: 15px;
+    border-left: 4px solid #4CAF50;
+    margin: 20px 0;
+    border-radius: 5px;
+}
+
+.api-result {
+    background: #f0f8ff;
+    padding: 15px;
+    margin: 10px 0;
+    border-left: 4px solid #2196F3;
+    border-radius: 5px;
+}
+
+.error {
+    background: #ffe8e8;
+    padding: 15px;
+    border-left: 4px solid #f44336;
+    margin: 10px 0;
+    border-radius: 5px;
+}
+
+.loading {
+    text-align: center;
+    padding: 20px;
+    color: #666;
+}
+
+pre {
+    background: #f8f8f8;
+    padding: 10px;
+    border-radius: 5px;
+    overflow-x: auto;
+    font-size: 12px;
+}
+
+button {
+    background: #4CAF50;
+    color: white;
+    padding: 12px 24px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+    margin: 10px 5px;
+}
+
+button:hover {
+    background: #45a049;
+}
+
+.sync-button {
+    background: #ff9800;
+}
+
+.sync-button:hover {
+    background: #f57c00;
+}
+
+.comparison {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.timing-info {
+    background: #fff3cd;
+    padding: 8px 12px;
+    border-radius: 4px;
+    border-left: 3px solid #ffc107;
+    margin: 5px 0;
+    font-size: 14px;
+}
+
+.timing-fast {
+    border-left-color: #28a745;
+    background: #d4edda;
+}
+
+.timing-medium {
+    border-left-color: #ffc107;
+    background: #fff3cd;
+}
+
+.timing-slow {
+    border-left-color: #dc3545;
+    background: #f8d7da;
+}
+
+.timing-summary {
+    background: #e9ecef;
+    padding: 15px;
+    border-radius: 5px;
+    margin: 15px 0;
+}
+
+.timing-stats {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    gap: 10px;
+    margin-top: 10px;
+}
+
+.stat-item {
+    background: white;
+    padding: 10px;
+    border-radius: 4px;
+    text-align: center;
+    border: 1px solid #dee2e6;
+}
+
+.js-results {
+    border: 2px solid #f7df1e;
+    border-radius: 10px;
+    padding: 20px;
+    margin: 20px 0;
+}
+
+.performance-comparison {
+    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+    color: white;
+    padding: 20px;
+    border-radius: 10px;
+    margin: 20px 0;
+}
+
+.performance-table {
+    width: 100%;
+    border-collapse: collapse;
+    margin: 20px 0;
+    background: white;
+    border-radius: 8px;
+    overflow: hidden;
+}
+
+.performance-table th,
+.performance-table td {
+    padding: 12px;
+    text-align: left;
+    border-bottom: 1px solid #ddd;
+    color: #333;
+}
+
+.performance-table th {
+    background: #f8f9fa;
+    font-weight: bold;
+}
+
+.winner {
+    background: #d4edda !important;
+    font-weight: bold;
+}
